requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
